Train part 1 (known bugs)
Chase, this has to do with how you coded the logs, so if you could take a look at it, I would appreciate it.

When you are moving left or rigth on the trainFlats which have the log tag, your character is stretched in the z dimention. If you get off the train by moving left or right off a flat, you die.